fix(logout): edit logout
